Small changes on storage embedded and cfg
